# API/Core rewrite: inspect dependencies without yarn|npm