… .travis.yml
… CHANGELOG.md
… CONTRIBUTING.md
… README.md
… package.json
… test/benchmark.js
… test/packages/huge/npm/serverless.js
… test/packages/huge/npm/serverless.yml
… test/packages/huge/yarn/serverless.js
… test/packages/huge/yarn/serverless.yml
… test/packages/individually/npm/serverless.js
… test/packages/individually/npm/serverless.yml
… test/packages/individually/yarn/serverless.js
… test/packages/individually/yarn/serverless.yml
… test/packages/simple/npm/serverless.js
… test/packages/simple/npm/serverless.yml
… test/packages/simple/yarn/serverless.js
… test/packages/simple/yarn/serverless.yml
… test/packages/webpack/npm/serverless.js
… test/packages/webpack/npm/serverless.yml
… test/packages/webpack/yarn/serverless.js
… test/packages/webpack/yarn/serverless.yml
… test/spec/glob.spec.js
… yarn.lock