## Tasks
- [x] Experiment with `depInclude` **after** `include`...
    - **Result**: Works for sure, but is a break from `serverless` packaging (e.g., what happens if you have the same configuration and remove Jetpack from plugins).

## Work

* Replace strategy of `yarn|npm install --production` in a temporary directory with better, faster production dependency inference via `inspectdep`.
* Change the API to take no `custom` options.
* Remove `lockfile` option from test / benchmark matrix.

## Faster timings!

| OS      | Scenario     | Type     | Time  |      vs Base |
| :------ | :----------- | :------- | ----: | -----------: |
| Mac     | simple       | jetpack  |  4338 | **-46.37 %** |
| Mac     | simple       | baseline |  8089 |              |
| Mac     | individually | jetpack  |  2964 | **-76.77 %** |
| Mac     | individually | baseline | 12760 |              |
| Mac     | huge         | jetpack  |  4524 | **-84.03 %** |
| Mac     | huge         | baseline | 28321 |              |
| Windows | simple       | jetpack  |  2890 | **-73.88 %** |
| Windows | simple       | baseline | 11063 |              |
| Windows | individually | jetpack  |  3110 | **-83.90 %** |
| Windows | individually | baseline | 19312 |              |
| Windows | huge         | jetpack  |  3500 | **-93.34 %** |
| Windows | huge         | baseline | 52548 |              |